Started to refactor path finding prior to add different pathfinding types
